test: align backup runner test files with source files
Move NamespaceBackupRunner tests into namespace_backup_runner_test.go
and reserve routine_backup_runner_test.go for RoutineBackupRunnerImpl.
