+++ pkg/service/routine_backup_runner_test.go
@@ -4,422 +4,104 @@ import (
 	"context"
 	"errors"
 	"log/slog"
-	"sync"
 	"testing"
 	"time"
 
 	"github.com/aerospike/aerospike-backup-service/v3/pkg/model"
-	"github.com/aerospike/aerospike-backup-service/v3/pkg/service/backupexecutor"
-	"github.com/aerospike/aerospike-backup-service/v3/pkg/util/optional"
+	"github.com/aerospike/aerospike-backup-service/v3/pkg/service/aerospike"
 	"github.com/aerospike/backup-go/models"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
-	"golang.org/x/sync/semaphore"
 )
 
-type testMocks struct {
-	ctrl           *gomock.Controller
-	backupExecutor *backupexecutor.MockBackup
-	backupHandler  *backupexecutor.MockBackupHandler
-	backendService *MockBackupReaderWriter
+type fakeNamespaceResolver struct {
+	namespaces []string
+	err        error
 }
 
-func initMocks(t *testing.T) (testMocks, NamespaceBackupRunner) {
-	t.Helper()
+func (f *fakeNamespaceResolver) ResolveNamespaces(
+	_ context.Context, _ *model.BackupRoutine, _ *slog.Logger,
+) ([]string, error) {
+	return f.namespaces, f.err
+}
+
+var _ aerospike.NamespaceResolver = (*fakeNamespaceResolver)(nil)
+
+func TestRoutineBackupRunner_Run_Success(t *testing.T) {
+	t.Parallel()
+
 	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
 
-	mockBackupExecutor := backupexecutor.NewMockBackup(ctrl)
-	mockBackupHandler := backupexecutor.NewMockBackupHandler(ctrl)
-	mockBackendService := NewMockBackupReaderWriter(ctrl)
+	nsRunner := NewMockNamespaceBackupRunner(ctrl)
+	resolver := &fakeNamespaceResolver{namespaces: []string{"ns1", "ns2"}}
 
-	executor := NewNamespaceBackupRunner(
-		mockBackupExecutor,
-		mockBackendService,
-		NewPathService(nil),
-	)
+	handler1 := NewMockCancelableBackupHandler(ctrl)
+	handler1.EXPECT().GetStats().Return(models.NewBackupStats()).AnyTimes()
+	handler2 := NewMockCancelableBackupHandler(ctrl)
+	handler2.EXPECT().GetStats().Return(models.NewBackupStats()).AnyTimes()
 
-	return testMocks{
-		ctrl:           ctrl,
-		backupExecutor: mockBackupExecutor,
-		backupHandler:  mockBackupHandler,
-		backendService: mockBackendService,
-	}, executor
-}
+	nsRunner.EXPECT().
+		Run(gomock.Any(), gomock.Any(), "ns1", gomock.Any(), gomock.Any(), gomock.Any()).
+		Return(handler1)
+	nsRunner.EXPECT().
+		Run(gomock.Any(), gomock.Any(), "ns2", gomock.Any(), gomock.Any(), gomock.Any()).
+		Return(handler2)
 
-func TestRun_SuccessfulFullBackup(t *testing.T) {
-	now := time.UnixMilli(123456789000)
-	backupFolder := "test-routine/backup/123456789000/data/test-ns"
-
-	mocks, runner := initMocks(t)
-	defer mocks.ctrl.Finish()
-
-	routine := &model.BackupRoutine{Name: routineName, SourceCluster: &model.AerospikeCluster{}}
-	timeBounds := model.TimeBounds{}
-	backupStats := models.NewBackupStats()
-	backupStats.TotalRecords.Store(100)
-	backupStats.ReadRecords.Add(50)
-
-	mocks.backupExecutor.EXPECT().
-		Run(gomock.Any(), routine, timeBounds, testNamespace, backupFolder, gomock.Any(), gomock.Any()).
-		Return(mocks.backupHandler, nil)
-
-	mocks.backupHandler.EXPECT().
-		Wait(gomock.Any()).
-		Return(nil)
-
-	mocks.backupHandler.EXPECT().
-		GetStats().
-		Return(backupStats)
-
-	mocks.backendService.EXPECT().
-		WriteBackupMetadata(gomock.Any(), routine, backupFolder, gomock.Any()).
-		DoAndReturn(func(_ context.Context, _ *model.BackupRoutine, _ string, metadata model.BackupMetadata) error {
-			// Check that metadata contains expected data
-			assert.Equal(t, testNamespace, metadata.Namespace)
-			assert.Equal(t, now, metadata.Created)
-			assert.Equal(t, time.Time{}, metadata.From) // full backup
-			assert.Equal(t, uint64(50), metadata.RecordCount)
-			return nil
-		})
-
-	spec := model.BackupRunSpec{Type: model.BackupTypeFull, StartTime: now, TimeBounds: timeBounds}
-	handler := runner.Run(t.Context(), routine, testNamespace, spec, nil, slog.Default())
-
-	require.NotNil(t, handler)
-	err := handler.Wait(t.Context())
-	require.NoError(t, err)
-}
-
-func TestSuccessfulIncrementalBackup(t *testing.T) {
-	now := time.UnixMilli(123456789000)
-	fromTime := time.UnixMilli(100000000000)
-	backupFolder := "test-routine/incremental/123456789000/data/test-ns"
-
-	mocks, runner := initMocks(t)
-	defer mocks.ctrl.Finish()
-
-	routine := &model.BackupRoutine{Name: routineName, SourceCluster: &model.AerospikeCluster{}}
-	timeBounds := model.TimeBounds{
-		FromTime: &fromTime,
-	}
-	backupStats := models.NewBackupStats()
-	backupStats.TotalRecords.Store(50)
-	backupStats.ReadRecords.Add(25)
-
-	mocks.backupExecutor.EXPECT().
-		Run(gomock.Any(), routine, timeBounds, testNamespace, backupFolder, gomock.Any(), gomock.Any()).
-		Return(mocks.backupHandler, nil)
-
-	mocks.backupHandler.EXPECT().
-		Wait(gomock.Any()).
-		Return(nil)
-
-	mocks.backupHandler.EXPECT().
-		GetStats().
-		Return(backupStats)
-
-	mocks.backendService.EXPECT().
-		WriteBackupMetadata(gomock.Any(), routine, backupFolder, gomock.Any()).
-		DoAndReturn(func(_ context.Context, _ *model.BackupRoutine, _ string, metadata model.BackupMetadata) error {
-			// Check that metadata contains expected data
-			assert.Equal(t, testNamespace, metadata.Namespace)
-			assert.Equal(t, now, metadata.Created)
-			assert.Equal(t, fromTime, metadata.From)
-			assert.Equal(t, uint64(25), metadata.RecordCount)
-			return nil
-		})
-
-	spec := model.BackupRunSpec{Type: model.BackupTypeIncremental, StartTime: now, TimeBounds: timeBounds}
-	handler := runner.Run(t.Context(), routine, testNamespace, spec, nil, slog.Default())
-
-	require.NotNil(t, handler)
-	err := handler.Wait(t.Context())
-	require.NoError(t, err)
-}
-
-func TestEmptyIncrementalBackup(t *testing.T) {
-	now := time.UnixMilli(123456789000)
-	fromTime := time.UnixMilli(100000000000)
-	backupFolder := "test-routine/incremental/123456789000/data/test-ns"
-
-	mocks, runner := initMocks(t)
-	defer mocks.ctrl.Finish()
-
-	routine := &model.BackupRoutine{Name: routineName, SourceCluster: &model.AerospikeCluster{}}
-	timeBounds := model.TimeBounds{
-		FromTime: &fromTime,
-	}
-
-	backupStats := models.NewBackupStats() // empty backup stats
-
-	mocks.backupExecutor.EXPECT().
-		Run(gomock.Any(), routine, timeBounds, testNamespace, backupFolder, gomock.Any(), gomock.Any()).
-		Return(mocks.backupHandler, nil)
-
-	mocks.backupHandler.EXPECT().
-		Wait(gomock.Any()).
-		Return(nil)
-
-	mocks.backupHandler.EXPECT().
-		GetStats().
-		Return(backupStats)
-
-	// No WriteBackupMetadata call expected for empty incremental backup.
-
-	spec := model.BackupRunSpec{Type: model.BackupTypeIncremental, StartTime: now, TimeBounds: timeBounds}
-	handler := runner.Run(t.Context(), routine, testNamespace, spec, nil, slog.Default())
-
-	require.NotNil(t, handler)
-	err := handler.Wait(t.Context())
-	require.NoError(t, err)
-}
-
-func TestBackupExecutorError(t *testing.T) {
-	now := time.UnixMilli(123456789000)
-	backupFolder := "test-routine/backup/123456789000/data/test-ns"
-
-	mocks, runner := initMocks(t)
-	defer mocks.ctrl.Finish()
-
+	runner := NewRoutineBackupRunner(nsRunner, resolver)
 	routine := &model.BackupRoutine{
-		Name:          routineName,
-		SourceCluster: &model.AerospikeCluster{},
-		BackupPolicy: &model.BackupPolicy{
-			RetryPolicy: &model.RetryPolicy{
-				MaxRetries: optional.Of(0),
-			},
-		},
+		Name:         "daily",
+		BackupPolicy: &model.BackupPolicy{},
 	}
-	timeBounds := model.TimeBounds{}
 
-	mocks.backupExecutor.EXPECT().
-		Run(gomock.Any(), routine, timeBounds, testNamespace, backupFolder, gomock.Any(), gomock.Any()).
-		Return(nil, errors.New("nsRunner error"))
-
-	// backup fails to start => nothing is written via backendService
-	spec := model.BackupRunSpec{Type: model.BackupTypeFull, StartTime: now, TimeBounds: timeBounds}
-	handler := runner.Run(t.Context(), routine, testNamespace, spec, nil, slog.Default())
-
-	require.NotNil(t, handler)
-	err := handler.Wait(t.Context())
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "failed to start backup")
-}
-
-func TestBackupHandlerError(t *testing.T) {
-	now := time.UnixMilli(123456789000)
-	backupFolder := "test-routine/backup/123456789000/data/test-ns"
-	timestampPath := "test-routine/backup/123456789000"
-
-	mocks, runner := initMocks(t)
-	defer mocks.ctrl.Finish()
-
-	routine := &model.BackupRoutine{
-		Name:          routineName,
-		SourceCluster: &model.AerospikeCluster{},
-		BackupPolicy: &model.BackupPolicy{
-			RetryPolicy: &model.RetryPolicy{
-				MaxRetries: optional.Of(0),
-			},
-		},
-	}
-	timeBounds := model.TimeBounds{}
-
-	mocks.backupExecutor.EXPECT().
-		Run(gomock.Any(), routine, timeBounds, testNamespace, backupFolder, gomock.Any(), gomock.Any()).
-		Return(mocks.backupHandler, nil)
-
-	mocks.backupHandler.EXPECT().
-		Wait(gomock.Any()).
-		Return(errors.New("handler error"))
-
-	mocks.backendService.EXPECT(). // Important: delete all backup files, including configuration
-					Delete(gomock.Any(), routine, timestampPath).
-					Return(nil)
-
-	spec := model.BackupRunSpec{Type: model.BackupTypeFull, StartTime: now, TimeBounds: timeBounds}
-	handler := runner.Run(t.Context(), routine, testNamespace, spec, nil, slog.Default())
-
-	require.NotNil(t, handler)
-	err := handler.Wait(t.Context())
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "backup failed")
-}
-
-func TestMetadataWriteError(t *testing.T) {
-	now := time.UnixMilli(123456789000)
-	backupFolder := "test-routine/backup/123456789000/data/test-ns"
-	timestampPath := "test-routine/backup/123456789000"
-
-	mocks, runner := initMocks(t)
-	defer mocks.ctrl.Finish()
-
-	routine := &model.BackupRoutine{
-		Name:          routineName,
-		SourceCluster: &model.AerospikeCluster{},
-		BackupPolicy: &model.BackupPolicy{
-			RetryPolicy: &model.RetryPolicy{
-				MaxRetries: optional.Of(0),
-			},
-		},
-	}
-	timeBounds := model.TimeBounds{}
-	backupStats := models.NewBackupStats()
-	backupStats.TotalRecords.Store(100)
-
-	metadataError := errors.New("metadata write error")
-	mocks.backupExecutor.EXPECT().
-		Run(gomock.Any(), routine, timeBounds, testNamespace, backupFolder, gomock.Any(), gomock.Any()).
-		Return(mocks.backupHandler, nil)
-
-	mocks.backupHandler.EXPECT().
-		Wait(gomock.Any()).
-		Return(nil)
-
-	mocks.backupHandler.EXPECT().
-		GetStats().
-		Return(backupStats)
-
-	mocks.backendService.EXPECT(). // backup succeeded, but failed to write metadata => delete all
-					WriteBackupMetadata(gomock.Any(), routine, backupFolder, gomock.Any()).
-					Return(metadataError)
-
-	mocks.backendService.EXPECT().
-		Delete(gomock.Any(), routine, timestampPath).
-		Return(nil)
-
-	spec := model.BackupRunSpec{Type: model.BackupTypeFull, StartTime: now, TimeBounds: timeBounds}
-	handler := runner.Run(t.Context(), routine, testNamespace, spec, nil, slog.Default())
-
-	require.NotNil(t, handler)
-	err := handler.Wait(t.Context())
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "failed to write backup metadata")
-}
-
-func TestRetryableBackupHandler_Cancel(t *testing.T) {
-	now := time.UnixMilli(123456789000)
-	backupFolder := "test-routine/backup/123456789000/data/test-ns"
-	timestampPath := "test-routine/backup/123456789000"
-
-	mocks, runner := initMocks(t)
-	defer mocks.ctrl.Finish()
-
-	routine := &model.BackupRoutine{Name: routineName, SourceCluster: &model.AerospikeCluster{}}
-	timeBounds := model.TimeBounds{}
-
-	var wg sync.WaitGroup
-	wg.Add(1)
-
-	// Set up minimal expectations to create a handler
-	mocks.backupExecutor.EXPECT().
-		Run(gomock.Any(), routine, timeBounds, testNamespace, backupFolder, gomock.Any(), gomock.Any()).
-		Return(mocks.backupHandler, nil)
-
-	mocks.backupHandler.EXPECT().
-		Wait(gomock.Any()).
-		DoAndReturn(func(ctx context.Context) error {
-			// Signal that Wait was called
-			wg.Done()
-			// This will hang until context is canceled
-			<-ctx.Done()
-			return ctx.Err()
-		})
-
-	mocks.backendService.EXPECT(). // backup canceled => delete all
-					Delete(gomock.Any(), routine, timestampPath).
-					Return(nil)
-
-	handler := runner.Run(
-		t.Context(),
-		routine,
-		testNamespace,
-		model.BackupRunSpec{Type: model.BackupTypeFull, StartTime: now, TimeBounds: timeBounds},
-		nil,
-		slog.Default(),
-	)
-
-	// Wait for the handler.Wait to be called
-	wg.Wait()
-
-	// Cancel the handler
-	handler.Cancel()
-
-	// Assert - the Wait method should eventually return with context.Canceled
-	err := handler.Wait(t.Context())
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "context canceled")
-}
-
-func TestRun_RetryRecalculatesPath(t *testing.T) {
-	startTime1 := time.UnixMilli(123456789000)
-	backupFolder1 := "test-routine/backup/123456789000/data/test-ns"
-	timestampPath1 := "test-routine/backup/123456789000"
-
-	mocks, runner := initMocks(t)
-	defer mocks.ctrl.Finish()
-
-	routine := &model.BackupRoutine{
-		Name:          routineName,
-		SourceCluster: &model.AerospikeCluster{},
-		BackupPolicy: &model.BackupPolicy{
-			RetryPolicy: &model.RetryPolicy{
-				MaxRetries:  optional.Of(1),
-				BaseTimeout: optional.Of(time.Millisecond),
-			},
-		},
-	}
-	timeBounds := model.TimeBounds{}
-
-	// First attempt fails
-	mocks.backupExecutor.EXPECT().
-		Run(gomock.Any(), routine, timeBounds, testNamespace, backupFolder1, gomock.Any(), gomock.Any()).
-		Return(mocks.backupHandler, nil)
-
-	mocks.backupHandler.EXPECT().
-		Wait(gomock.Any()).
-		Return(errors.New("handler error"))
-
-	mocks.backendService.EXPECT().
-		Delete(gomock.Any(), routine, timestampPath1).
-		Return(nil)
-
-	// Second attempt succeeds with new folder
-	// We use DoAndReturn to capture the startTime updated in OnRetry
-	mocks.backupExecutor.EXPECT().
-		Run(gomock.Any(), routine, timeBounds, testNamespace, gomock.Any(), gomock.Any(), gomock.Any()).
-		DoAndReturn(func(
-			_ context.Context,
-			_ *model.BackupRoutine,
-			_ model.TimeBounds,
-			_ string,
-			backupFolder string,
-			_ *semaphore.Weighted,
-			_ *slog.Logger,
-		) (backupexecutor.BackupHandler, error) {
-			assert.NotEqual(t, backupFolder1, backupFolder)
-			assert.Contains(t, backupFolder, "test-routine/backup/")
-			return mocks.backupHandler, nil
-		})
-
-	mocks.backupHandler.EXPECT().
-		Wait(gomock.Any()).
-		Return(nil)
-
-	backupStats := models.NewBackupStats()
-	mocks.backupHandler.EXPECT().
-		GetStats().
-		Return(backupStats)
-
-	mocks.backendService.EXPECT().
-		WriteBackupMetadata(gomock.Any(), routine, gomock.Any(), gomock.Any()).
-		Return(nil)
-
-	spec := model.BackupRunSpec{Type: model.BackupTypeFull, StartTime: startTime1, TimeBounds: timeBounds}
-	handler := runner.Run(t.Context(), routine, testNamespace, spec, nil, slog.Default())
-
-	require.NotNil(t, handler)
-	err := handler.Wait(t.Context())
+	op, err := runner.Run(t.Context(), routine, model.BackupRunSpec{Type: model.BackupTypeFull}, slog.Default())
 	require.NoError(t, err)
+	require.NotNil(t, op)
+	assert.Len(t, op.handlers, 2)
+}
+
+func TestRoutineBackupRunner_Run_ResolverError(t *testing.T) {
+	t.Parallel()
+
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	nsRunner := NewMockNamespaceBackupRunner(ctrl)
+	resolver := &fakeNamespaceResolver{err: errors.New("resolve failed")}
+
+	runner := NewRoutineBackupRunner(nsRunner, resolver)
+	routine := &model.BackupRoutine{Name: "daily", BackupPolicy: &model.BackupPolicy{}}
+
+	op, err := runner.Run(t.Context(), routine, model.BackupRunSpec{Type: model.BackupTypeFull}, slog.Default())
+	require.Error(t, err)
+	assert.Contains(t, err.Error(), "resolve failed")
+	assert.Nil(t, op)
+}
+
+func TestRoutineBackupRunner_Run_WaitUntilStartedTimesOut(t *testing.T) {
+	t.Parallel()
+
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	nsRunner := NewMockNamespaceBackupRunner(ctrl)
+	resolver := &fakeNamespaceResolver{namespaces: []string{"ns1"}}
+
+	handler := NewMockCancelableBackupHandler(ctrl)
+	handler.EXPECT().GetStats().Return(nil).AnyTimes()
+
+	nsRunner.EXPECT().
+		Run(gomock.Any(), gomock.Any(), "ns1", gomock.Any(), gomock.Any(), gomock.Any()).
+		Return(handler)
+
+	runner := NewRoutineBackupRunner(nsRunner, resolver)
+	routine := &model.BackupRoutine{Name: "daily", BackupPolicy: &model.BackupPolicy{}}
+
+	ctx, cancel := context.WithTimeout(t.Context(), 30*time.Millisecond)
+	defer cancel()
+
+	op, err := runner.Run(ctx, routine, model.BackupRunSpec{Type: model.BackupTypeFull}, slog.Default())
+	require.Error(t, err)
+	assert.Nil(t, op)
 }
